requirements.txt
pandas
plotly
matplotlib
streamlit_lottie
streamlit_option_menu